refactor(compositions): factoriser renderPlayerItem et logique burnedTeam
- Créer composant AvailablePlayerItem pour mutualiser l'affichage des joueurs
- Créer utilitaire getBurnedTeamNumber pour centraliser le calcul du brûlage
- Créer type DiscordMember dans @/types/discord
- Supprimer code dupliqué dans compositions/page.tsx et defaults/page.tsx
- Corriger imports inutilisés et warnings ESLint
