+++ src/types/discord.ts
@@ -1,0 +1,13 @@
+export interface DiscordMember {
+  id: string;
+  username: string;
+  displayName: string;
+  avatar?: string | null;
+}
+
+export interface DiscordChannel {
+  id: string;
+  name: string;
+  type: number;
+}
+
